refactor(audit): C1-2 move the 7 stage helpers into core/pipeline_stages.py
impulcifer.py shrinks to version + main() + CLI (912 -> 260 lines) and
keeps identity re-exports; BRIRPipeline.run() now imports the stages from
core, so the impulcifer <-> core.pipeline import cycle is gone. Moved
bodies are AST-identical except two intended edits: the source-tree data/
fallback resolves via the repo root (module moved into core/) and
core.hrir._get_center_value is now public get_center_value (F036).

Refs #138
